Add off-by-one mutation
As suggeested on https://github.com/llogiq/mutagen/issues/5, it would be
nice to mutate numeric constants adding or substracting one unit.

At the moment, this is very conservative on the limits to add or
substract on default integer constants. I've did it like this to avoid
compilation issue, as we do not have information regarding the type of
the expression. So, in the default case, it subtracts one only if it's >
0 and it only adds one if the value is < 256.

If the constant has a type annotation (for example, 0i32), I adjust the
minimum and the maximum depending on the type hint. Note that it does
not check u128, as it's only available on nightly.

Finally, it chains the mutations in order to record only the needed
mutations. So, if the code finds a "0" on an unsigned type, it will only
record the "add one" mutation, but not the "sub one".