Check for negative numbers
As Rust AST does not track negative literals on ExprKind::Lit, the code
has been changed in order to check for Unary operations which contains a
negative operand and then the literal. The code has been factored in
order to allow to entry points: one for negative literals and one for
positive literals.