Add testing framework
Adding a minimal testing framework, which will allow to test the plugin
and the mutations generations.

The rationale behind is create integration tests with one or more blocks
with the #[mutate] annotation. As the test will be compiled before
executing, when the tests will be excuted, we will be able to find the
`mutations.txt` file generated for this specific integration test.

Then, we can assert one or more messages with should be located on some
specific line. Note that we should be able to check also that some
mutation is not generated (for example: `assert_eq!(false,
checker.has("replacing _ || _ with false", "12"))`, will check that
this mutation is not being applied on line 12).

The tests will be fragile, as they will fail if:
1- Mutation output format is changed
2- If the spans of the testing example is changed
3- If any message is changed
4- Will fail if integration tests are executed concurrently (as there
will be issues with writting to the same file, reseting it, ...)

Also, checking which mutations has been generated is not as performant
as could be, but I think that it should be neglibile on this stage and
that may be improved on the future.