Add compiletest support
Adding compiletest library and adding an example for eq and ne mutations
that are applied properly.

On the other hand, I'm adding some tests that currently are failing (on
the compile-fail folder) that we should fix at some time. I think it
may be useful to keep track of the issues pending to be fixed and it
will help to not add regressions of known patterns that caused issues on
the past