Run only the tests which contains some mutation
Attempt to fix issue: https://github.com/llogiq/mutagen/issues/28

As a first step, all the mutations are being instrumented and are
calling `::mutagen::report_coverage()`. This method, if an environment
variables is set (`MUTAGEN_COVERAGE`), the first time it will be called,
it will write to a specific file that some mutation has been hit.

On the other side, `cargo-mutagen` will check all the tests that the
binary contains, and will execute all the tests individually, with the
mentioned env var. After executing any of the tests, we will check if
the file has been created. If it exists, we know that the executed test
contains code that it's mutated.

Finally, when running mutations, we will run only the tests that
contains mutations (unfortunatelly, we will need to run the test
one-by-one, which means that we will spawn a process per test, instead
of a process per testsuite). Then, coverage flag it's only useful when
the ratio of tests which executes mutated code is low.

To test with coverage, after updating to the new version of the plugin
on the target project, and after reinstalling the runner:

```
cargo mutagen -- --coverage
```

Pending tasks would be:
- Improve the "ping" system with something better than a filename
- Maybe we can include the mutation number to `report_coverage` method,
  in order to know which are the mutations that took place on the given
test. If we do, we can filter more accuretly which tests should be
executed on each mutation
- Maybe we can think about use clap to parse arguments, add subcommands,
  ...