System to avoid redundant mutations
The system will have a stack for all the generated mutation types: it
will contain an entry for each recursive call to `fold_expr`. When we
are performing a mutation, we are able to impose some restrictions to
the direct sub expressions.

So, if we are mutating an `&&`, we can force to left and
right expressions to do not emit a `replace with false` mutation, as we
will already mutate this and to `false`, and that would lead to a
redundant mutation.